+++ examples/dsp/mod.rs
@@ -1,2 +1,0 @@
-pub mod osc;
-pub mod wavetable;