+++ CHANGELOG.md
@@ -6,6 +6,9 @@ backwards compatibility.
 
 ## Unreleased
 
+* Breaking: Audio processing callback is now passed to
+  `handle_interrupt_dma1_str1` instead of `spawn`.
+
 ## 0.5.0
 
 * Add an example of OLED display configuration.